+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
+### Changed
+* Python >=3.11 required
 
 ## [3.3.1](https://pypi.org/project/lupyne/3.3.1/) - 2025-11-02
 ### Changed